fixed upgrading errors
fixed errors caused by upgrading the project to NextJS 15

--- NOTES ---
`noop-ai-web\src\components\charts\geography-map.tsx`

this file is commented out because the library use in this file `react-simple-maps` has dependency issue. it requires react 18 or lover. for now we don't need this actually. for now let's not use it. later if needed let's find a another library that supports react 19.

- the version we used 3.0.0 was the latest and doesn't seems to be updated in 2 years

------------------

added some `INFO`/`TODO` comments for the changes mase in some places